2HANDS/requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.4.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
click==8.1.7
contourpy==1.1.1
controlnet-aux==0.0.3
cycler==0.12.1
cython==3.0.10
diffusers==0.27.2
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.1
fastapi==0.108.0
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
funcy==1.18
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
gradio==4.21.0
gradio-client==0.12.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hickle==5.0.3
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
imageio==2.34.2
imageio-ffmpeg==0.5.1
imgaug==0.2.6
importlib-metadata==7.2.1
importlib-resources==6.4.0
iopaint==1.3.3
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-loader==0.4
lightning-utilities==0.11.5
loguru==0.7.2
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.5
packaging==24.1
pandas==2.0.3
peft==0.7.1
piexif==1.1.3
pillow==9.5.0
pkgutil-resolve-name==1.3.10
progressbar2==4.4.2
protobuf==5.27.1
psutil==6.0.0
pudb==2019.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.4
pydantic-core==2.18.4
pydeck==0.9.1
pydeprecate==0.3.1
pydub==0.25.1
pygments==2.18.0
pyparsing==3.1.2
pyside6-essentials==6.6.3.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.7.2
python-utils==3.8.2
pytorch-lightning==1.4.2
pytz==2024.1
pywavelets==1.4.1
pyyaml==6.0.1
referencing==0.35.1
regex==2024.5.15
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.10
safetensors==0.4.3
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
shiboken6==6.6.3.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.32.0.post1
streamlit==1.36.0
sympy==1.12.1
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
tifffile==2023.7.10
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typer==0.12.3
typer-config==1.4.0
tzdata==2024.1
urwid==2.6.15
uvicorn==0.30.1
watchdog==4.0.1
wcwidth==0.2.13
websockets==11.0.3
werkzeug==3.0.3
wsproto==1.2.0
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
